remove BgMusic folder and remove all logic music to ActionButtons
